Implement --passtos for tagged ethernet frames
This patch makes it possible to use the --passtos option with 802.1Q tagged
ethernet frames.

sf.net tracker:
<https://sourceforge.net/tracker/?func=detail&aid=2829878&group_id=48978&atid=454721>

Discussed on the IRC meeting March 4, 2010 in #openvpn-discussions.
<http://thread.gmane.org/gmane.network.openvpn.devel/3242>

* [PATCH] Implement --passtos for tagged ethernet frames
  - http://thread.gmane.org/gmane.network.openvpn.devel/3165
  - decided to ask for testing help in the mailinglists

This patch will need to be tested in it's own feat_passtos branch.  When testing
has been acceped by the community, it will be merged into the 'allmerged' branch.

Signed-off-by: David Sommerseth <dazo@users.sourceforge.net>