vlan: Add global --vlan-accept option
This patch adds the new "--vlan-accept mode" option.  The option is valid
in server mode.  It is set to "raw" by default.

If the option is set to "raw", OpenVPN will operate as before. Frames may
be tagged or untagged and OpenVPN couldn't care less.

If the option is set to "tagged", OpenVPN will assume the tap device
to be in tagged mode, i.e. packets coming in on the device are VLAN-tagged.
Packets going out through the device will be VLAN-tagged likewise. Untagged
packets coming in on the tap device will be dropped.

If the option is set to "untagged", openvpn will assume the packets to be
untagged or priority-tagged. Incoming VLAN-tagged packets will be dropped.

If the option is set to "all", openvpn will accept both VLAN-tagged packets
and untagged or priority-tagged packets.

The option has no immediate effect yet, but will be used by later patches.

Signed-off-by: Fabian Knittel <fabian.knittel@avona.com>