vlan: Add per-client --vlan-pvid option
This patch adds the new "--vlan-pvid" integer option.  The option is valid
in server mode and can be set in a client context (e.g. from the client-connect
hook).  It defaults to 1.

The value indicates which VID (VLAN identifier) to associate with a client
or (in global context) with the tap device.
It also serves as the Port VLAN Identifier, i.e. the VID to use for incoming
and outgoing untagged or priority-tagged frames.

The client will only receive packets which belong to the same VLAN.

The option has no immediate effect yet, but will be used by later patches.