vlan: Prepend and remove VLAN identifiers on outgoing and incoming frames
This patch adds parsing of the IEEE 802.1Q headers for incoming and outgoing
ethernet frames.

For frames coming in from the tap interface, the 802.1Q header is parsed and
translated into a regular Ethernet II header.

For frames where the Priority Code Point (PCP) is non-zero, the 802.1Q tagging
isn't removed. Instead, the VID is zeroed out, turning the frame into a
priority frame.

For frames going out over the tap interface, the Ethernet II header is
replaced with a 802.1Q-based header.  PCP and CFI are permanently set to 0,
unless the frame was a priority frame, in which case the values are copied
over.  The VID is set according to the instance's vlan_pvid value (also see
patch adding the "--vlan-pvid" option).

The patch also introduces the log level D_VLAN_DEBUG for debug messages
regarding tagging/untagging.