vlan: split out '--vlan-accept raw' to '--vlan-tagging' boolean switch
As '--vlan-accept raw' is a global behavioural modifier, it makes more sense
to break it out into a global '--vlan-tagging' boolean switch. (Which
matches the approach chosen in the initial patch-set.)

'--vlan-accept tagged|untagged|all' now only specifies the behaviour of the
tap device, while '--vlan-tagging' specifies whether OpenVPN should take a
closer look at the tagging of packets.

The patch is only a configuration option change.  The behaviour should
remain the same.

  New                                       Old
  ------------------------------------------------------------------

  not specifying anything is equivalent to not specifying anything

  not specifying anything is equivalent to '--vlan-accept raw'

  '--vlan-tagging' +
  '--vlan-accept tagged'  is equivalent to '--vlan-accept tagged'

  '--vlan-tagging' +
  '--vlan-accept untagged' is equivalent to '--vlan-accept untagged'

  '--vlan-tagging'        is equivalent to '--vlan-accept all'

  '--vlan-tagging' +
  '--vlan-accept all'     is equivalent to '--vlan-accept all'

Signed-off-by: Fabian Knittel <fabian.knittel@avona.com>