vlan: don't create priority-tagged frames
This patch changes the behaviour of tagged frames that contain both VLAN and
priority (PCP) information.  Previously, when stripping the VLAN tagging, the
priority information would be kept, causing priority-tagged frames to be sent
to clients with no tagging support.  The new behaviour is to completely drop
the tagging.

Previously some tagging-aware appliances that set both the VID and PCP fields
would cause OpenVPN to forward prio-tagged frames to clients, which wouldn't
understand and therefore drop them.

The only case where the tagging is kept is when the frame isn't vlan-tagged and
only prio-tagged.  In that case the sender clearly wanted to send a tagged
frame on an otherwise untagged network and we honour that decision by not
fiddling with it.

Signed-off-by: Fabian Knittel <fabian.knittel@lettink.de>